Pass HTML props to child <Tr>s
Since we're parsing data out of all the React node children that are
passed to the Table, we lose all of the metadata (such as props) that
are applied to those children. In this case, the data object is
actually a meta-object containing the data and the props, and signified
by a `__reactableMeta: true` property.

This probably still needs to be done for the Tds as well.

Fixes #103