Allow user defined Thead within Table
Since displayName is only available via a Components constructor
and not a ReactElement, this is a work around to allow you to add
displayName="thead" as a prop on your Thead element so Reactable
can differentiate. This should probably be a thing baked into React.

Also with this change it expects you to have Thead > Tr > Th's